refactor: move CrossIconButton into SearchBox since the overrideName means this component can't be re-used outside of the SearchBox anyway
