--- app/components/ui/molecules/CrossIconButton.md
@@ -1,5 +1,0 @@
-Clickable cross icon
-
-```js
-<CrossIconButton onClick={() => console.log('clicked')} />
-```
